SONARJAVA-2549 Move S2293 test files
