+++ java-checks-test-sources/src/main/java/checks/DiamondOperatorCheck_java_7.java
@@ -1,11 +1,12 @@
+package checks;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.function.Function;
 
-class A {
+class DiamondOperatorCheck_java_7 {
   List<Object> myList1 = new ArrayList<>(); // Compliant
   List<Object> myList2 = new ArrayList<Object>(); // Noncompliant [[sc=39;ec=47]] {{Replace the type specification in this constructor call with the diamond operator ("<>").}}
 
@@ -20,8 +21,8 @@ class A {
     Map<String,List<Integer>> map2 = new HashMap<String,List<Integer>>(); // Noncompliant [[sc=49;ec=71]]
 
     List myOtherList = new ArrayList<Object>(); // Compliant
-    new A().myList1 = new ArrayList<>(); // Compliant
-    new A().myList1 = new ArrayList<Object>(); // Noncompliant [[sc=36;ec=44]]
+    new DiamondOperatorCheck_java_7().myList1 = new ArrayList<>(); // Compliant
+    new DiamondOperatorCheck_java_7().myList1 = new ArrayList<Object>(); // Noncompliant [[sc=62;ec=70]]
 
     List<Object>[] myArrayOfList = new List[10];
     myArrayOfList[0] = new ArrayList<>(); // Compliant
@@ -33,9 +34,9 @@ class A {
     ((List<Object>) new ArrayList<>()).isEmpty(); // Compliant
 
     Iterator<Object> iterator = new Iterator<Object>() { // Compliant - anonymous classes requires to be typed
+      @Override public Object next() { return null; }
+      @Override public boolean hasNext() { return false; }
     };
-
-    MyUnknownVariable = new ArrayList<Object>(); // Compliant
 
     Object data = new List[10];
     ((List[])data)[2] = new ArrayList<String>();
@@ -44,30 +45,12 @@ class A {
   List<Object> qix(boolean test) {
     List<Object> myList = test ?
       new ArrayList<>() : // Compliant
-        new ArrayList<Object>(); // Noncompliant [[sc=22;ec=30]]
-    
+        new ArrayList<Object>(); // Compliant - using diamond operator only works with java 8
+
     myList = new ArrayList<>(test ? new ArrayList<>() : new ArrayList<Object>(5)); // Compliant
     if (test) {
       return new ArrayList<>(); // Compliant
     }
     return new ArrayList<Object>(); // Noncompliant [[sc=25;ec=33]]
   }
-
-  public void assignmentOnMethodInvocation() {
-    this.bar()[0] = new GenericClass<String, Integer, Integer>();
-  }
-
-  public MyInterface[] bar() {
-    return new MyInterface[1];
-  }
-
-  interface MyInterface { }
-  static class GenericClass<X, Y, Z> implements MyInterface { }
-}
-
-enum SonarProblemCase {
-  MY_BUCKET(rp -> { return new ArrayList<Integer>(); }),
-  MY_BUCKET2(rp -> new ArrayList<Integer>());
-
-  SonarProblemCase(Function<Object, Object> bucketer) {}
 }
