+++ java-checks-test-sources/src/main/java/checks/DiamondOperatorCheck_no_version.java
@@ -1,12 +1,14 @@
+package checks;
+
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
-class A {
+class DiamonOperatorCheck_no_version {
   List<Object> myList1 = new ArrayList<>(); // Compliant
-  List<Object> myList2 = new ArrayList<Object>(); // Noncompliant [[sc=39;ec=47]] {{Replace the type specification in this constructor call with the diamond operator ("<>").}}
+  List<Object> myList2 = new ArrayList<Object>(); // Noncompliant [[sc=39;ec=47]] {{Replace the type specification in this constructor call with the diamond operator ("<>"). (sonar.java.source not set. Assuming 7 or greater.)}}
 
   void foo() {
     List<Object> myList;
@@ -19,8 +21,8 @@ class A {
     Map<String,List<Integer>> map2 = new HashMap<String,List<Integer>>(); // Noncompliant [[sc=49;ec=71]]
 
     List myOtherList = new ArrayList<Object>(); // Compliant
-    new A().myList1 = new ArrayList<>(); // Compliant
-    new A().myList1 = new ArrayList<Object>(); // Noncompliant [[sc=36;ec=44]]
+    new DiamonOperatorCheck_no_version().myList1 = new ArrayList<>(); // Compliant
+    new DiamonOperatorCheck_no_version().myList1 = new ArrayList<Object>(); // Noncompliant [[sc=65;ec=73]]
 
     List<Object>[] myArrayOfList = new List[10];
     myArrayOfList[0] = new ArrayList<>(); // Compliant
@@ -32,9 +34,9 @@ class A {
     ((List<Object>) new ArrayList<>()).isEmpty(); // Compliant
 
     Iterator<Object> iterator = new Iterator<Object>() { // Compliant - anonymous classes requires to be typed
+      @Override public Object next() { return null; }
+      @Override public boolean hasNext() { return false; }
     };
-
-    MyUnknownVariable = new ArrayList<Object>(); // Compliant
 
     Object data = new List[10];
     ((List[])data)[2] = new ArrayList<String>();
@@ -43,8 +45,8 @@ class A {
   List<Object> qix(boolean test) {
     List<Object> myList = test ?
       new ArrayList<>() : // Compliant
-        new ArrayList<Object>(); // Compliant - using diamond operator only works with java 8
-    
+        new ArrayList<Object>(); // Noncompliant [[sc=22;ec=30]]
+
     myList = new ArrayList<>(test ? new ArrayList<>() : new ArrayList<Object>(5)); // Compliant
     if (test) {
       return new ArrayList<>(); // Compliant
